workers/score-equator/requirements.txt
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
google-cloud-bigquery>=3.17.0
google-cloud-firestore>=2.14.0
google-cloud-pubsub>=2.19.0
